Lower CPU level for init routines
