+++ config.make.in
@@ -28,6 +28,7 @@ CXX := @CXX@
 LD := @LD@
 AR := @AR@
 ARFLAGS := @ARFLAGS@
+CFLAGS_INIT := @CFLAGS_INIT@ "@CFLAGS_QUOTE_INIT@"
 CFLAGS := @CFLAGS@ "@CFLAGS_QUOTE@"
 CFLAGS_DEBUG := @CFLAGS_DEBUG@
 CFLAGS_SHARED := @CFLAGS_SHARED@
