+++ zdnn/zdnn.c
@@ -30,91 +30,6 @@
 #include <ihapsa.h>
 #endif
 
-#ifdef __MVS__
-#pragma export(zdnn_init)
-#endif
-
-// global variables, set by zdnn_init() via environment vars
-log_levels log_level = LOGLEVEL_ERROR; // log level (see enum log_levels)
-bool precheck_enabled = false; // enables tensor pre-check before invoking NNPA
-uint32_t status_diag = STATUS_DIAG_NOT_SET; // diagnostic info when status = X
-char log_module[LOGMODULE_SIZE] = "\0";
-
-/// Initialize the zDNN library and issue NNPA-QAF to the hardware.  Needs to be
-/// invoked at least once during the lifetime of the application, either
-/// manually or automatically via DLL-load initializer class.
-///
-/// \return None
-///
-INIT_FUNCTION_ATTRS
-void zdnn_init() {
-
-  char *ptr, *endptr;
-
-  if ((ptr = getenv(ENVVAR_LOGLEVEL))) {
-    if (!strcasecmp("off", ptr)) {
-      log_level = LOGLEVEL_OFF;
-    }
-
-    if (!strcasecmp("fatal", ptr)) {
-      log_level = LOGLEVEL_FATAL;
-    }
-
-    if (!strcasecmp("error", ptr)) {
-      log_level = LOGLEVEL_ERROR;
-    }
-
-    if (!strcasecmp("warn", ptr)) {
-      log_level = LOGLEVEL_WARN;
-    }
-
-    if (!strcasecmp("info", ptr)) {
-      log_level = LOGLEVEL_INFO;
-    }
-
-    if (!strcasecmp("debug", ptr)) {
-      log_level = LOGLEVEL_DEBUG;
-    }
-
-    if (!strcasecmp("trace", ptr)) {
-      log_level = LOGLEVEL_TRACE;
-    }
-  }
-
-  if ((ptr = getenv(ENVVAR_ENABLE_PRECHECK))) {
-    precheck_enabled = !strcasecmp("true", ptr);
-  }
-
-  if ((ptr = getenv(ENVVAR_STATUS_DIAG))) {
-
-    uint32_t val;
-
-    // if it's prefixed with "0x"/"0X" then treat it as hex string
-    if (strstr(ptr, "0x") == ptr || strstr(ptr, "0X") == ptr) {
-      val = (uint32_t)strtol((ptr + 2), &endptr, 16);
-    } else {
-      val = (uint32_t)strtol(ptr, &endptr, 10);
-    }
-
-    if (endptr == ptr + strlen(ptr)) {
-      status_diag = val;
-    }
-  }
-
-  if ((ptr = getenv(ENVVAR_LOGMODULE))) {
-    strncpy(log_module, ptr, LOGMODULE_SIZE - 1);
-  }
-
-  /* Exit silently if there is no NNPA facility installed.  Explicit
-     invocations of functions requiring NNPA will result in an
-     error.  */
-#ifndef ZDNN_CONFIG_NO_NNPA
-  if (zdnn_is_nnpa_installed() == false)
-    return;
-#endif
-  zdnn_refresh_nnpa_query_result();
-}
-
 /// Fill in NNPA tensor descriptor
 ///
 /// \param[out] descriptor pointer to an nnpa_parameter_block's descriptor
